Update config-v1.1.xml.php
Modified to allow server hostname to be configured in config.php or fetched from the server, the latter gives a single hostname back which pins email to a single server. In my case this is the ISPConfig master server which kills mail redundancy for users. Also added the option to append the email address to the Service name. Useful if users have multiple accounts on the same domain and/or server.